requirements.txt
certifi==2018.8.24
chardet==3.0.4
coverage==4.5.4
defusedxml==0.5.0
Django==2.2.4
django-allauth==0.37.1
django-cors-headers==2.4.0
django-filters==0.2.1
django-rest-auth==0.9.3
djangorestframework==3.9.3
drf-extensions==0.5.0
idna==2.7
oauthlib==2.1.0
python3-openid==3.1.0
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
sqlparse==0.3.0
urllib3==1.23
wincertstore==0.2
